+++ CHANGELOG.md
@@ -4,6 +4,10 @@
 
 ### Features
 
+- allow `processAttribute` hook to yield multiple attributes, typically used
+  when adding aliased attributes such as `:class`. By adding both the alias and
+  the original the latter can be validated as well (e.g. `no-dup-attr` can
+  trigger for multiple uses of `:class`). (fixes #36)
 - allow setting hooks when using `HtmlValidate.validateString`, makes it easier
   to write tests which requires hooks, e.g. processing attributes.